# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  auth.LoginRequset:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        minLength: 2
        type: string
      last_name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  cart.AddCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  cart.CartListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      total_price:
        type: number
    type: object
  cart.UpdateCartItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  http.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: integer
    type: object
  models.CartItem:
    properties:
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      user_id:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password_hash:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для e-commerce платформи з управлінням продуктами, кошиком та замовленнями
  title: E-Commerce API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Логін користувача та повертає токени доступу
      parameters:
      - description: Дані для логіну
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequset'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Логін користувача
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Оновлює access token за допомогою refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Оновлення токена доступу
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Реєструє нового користувача та повертає токени доступу
      parameters:
      - description: Дані для реєстрації
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Реєстрація нового користувача
      tags:
      - auth
  /cart:
    delete:
      consumes:
      - application/json
      description: Видаляє всі товари з кошика користувача
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Очищає кошик
      tags:
      - cart
    get:
      consumes:
      - application/json
      description: Повертає всі товари користувача в кошику
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cart.CartListResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Повертає список товарів у кошику
      tags:
      - cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Додає новий товар у кошик користувача
      parameters:
      - description: Товар для додавання
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/cart.AddCartItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CartItem'
        "400":
          description: Invalid request body or quantity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Додає товар у кошик
      tags:
      - cart
  /cart/items/{id}:
    delete:
      consumes:
      - application/json
      description: Видаляє конкретний товар з кошика користувача
      parameters:
      - description: ID товару у кошику
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid item ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Видаляє товар з кошика
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Змінює кількість або товар у кошику користувача
      parameters:
      - description: ID товару у кошику
        in: path
        name: id
        required: true
        type: string
      - description: Оновлені дані товару
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/cart.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CartItem'
        "400":
          description: Invalid request body or quantity
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: User not authorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Оновлює товар у кошику
      tags:
      - cart
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"